Move model parameters code to `model::parameters` submodule
